Added text and fixed list constructor

+++ TaxJar/Classes/TaxJarTaxOrderResponse.cs
@@ -12,47 +12,47 @@ namespace TaxJar.Classes
 
     public class TaxJarLineItemResponse    {
         public string id { get; set; } 
-        public int taxable_amount { get; set; } 
-        public double tax_collectable { get; set; } 
-        public double combined_tax_rate { get; set; } 
-        public int state_taxable_amount { get; set; } 
-        public double state_sales_tax_rate { get; set; } 
-        public double state_amount { get; set; } 
-        public int county_taxable_amount { get; set; } 
-        public double county_tax_rate { get; set; } 
-        public double county_amount { get; set; } 
-        public int city_taxable_amount { get; set; } 
-        public int city_tax_rate { get; set; } 
-        public int city_amount { get; set; } 
-        public int special_district_taxable_amount { get; set; } 
-        public double special_tax_rate { get; set; } 
-        public double special_district_amount { get; set; } 
+        public decimal taxable_amount { get; set; } 
+        public decimal tax_collectable { get; set; } 
+        public decimal combined_tax_rate { get; set; } 
+        public decimal state_taxable_amount { get; set; } 
+        public decimal state_sales_tax_rate { get; set; } 
+        public decimal state_amount { get; set; } 
+        public decimal county_taxable_amount { get; set; } 
+        public decimal county_tax_rate { get; set; } 
+        public decimal county_amount { get; set; } 
+        public decimal city_taxable_amount { get; set; } 
+        public decimal city_tax_rate { get; set; } 
+        public decimal city_amount { get; set; } 
+        public decimal special_district_taxable_amount { get; set; } 
+        public decimal special_tax_rate { get; set; } 
+        public decimal special_district_amount { get; set; } 
     }
 
     public class TaxJarBreakdownResponse    {
-        public int taxable_amount { get; set; } 
-        public double tax_collectable { get; set; } 
-        public double combined_tax_rate { get; set; } 
-        public int state_taxable_amount { get; set; } 
-        public double state_tax_rate { get; set; } 
-        public double state_tax_collectable { get; set; } 
-        public int county_taxable_amount { get; set; } 
-        public double county_tax_rate { get; set; } 
-        public double county_tax_collectable { get; set; } 
-        public int city_taxable_amount { get; set; } 
-        public int city_tax_rate { get; set; } 
-        public int city_tax_collectable { get; set; } 
-        public int special_district_taxable_amount { get; set; } 
-        public double special_tax_rate { get; set; } 
-        public double special_district_tax_collectable { get; set; } 
+        public decimal taxable_amount { get; set; } 
+        public decimal tax_collectable { get; set; } 
+        public decimal combined_tax_rate { get; set; } 
+        public decimal state_taxable_amount { get; set; } 
+        public decimal state_tax_rate { get; set; } 
+        public decimal state_tax_collectable { get; set; } 
+        public decimal county_taxable_amount { get; set; } 
+        public decimal county_tax_rate { get; set; } 
+        public decimal county_tax_collectable { get; set; } 
+        public decimal city_taxable_amount { get; set; } 
+        public decimal city_tax_rate { get; set; } 
+        public decimal city_tax_collectable { get; set; } 
+        public decimal special_district_taxable_amount { get; set; } 
+        public decimal special_tax_rate { get; set; } 
+        public decimal special_district_tax_collectable { get; set; } 
         public List<TaxJarLineItemResponse> line_items { get; set; } 
     }
 
     public class TaxJarTaxResponse    {
-        public double order_total_amount { get; set; } 
-        public double shipping { get; set; } 
+        public decimal order_total_amount { get; set; } 
+        public decimal shipping { get; set; } 
         public decimal taxable_amount { get; set; } 
-        public double amount_to_collect { get; set; } 
+        public decimal amount_to_collect { get; set; } 
         public decimal rate { get; set; } 
         public bool has_nexus { get; set; } 
         public bool freight_taxable { get; set; }
